requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
backoff==2.2.1
    # via gql
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via openai
exceptiongroup==1.2.1
    # via anyio
gql==3.4.1
    # via omnivoreql
graphql-core==3.2.3
    # via gql
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.0.5
    # via yarl
omnivoreql==0.2.1
openai==1.23.3
pydantic==2.7.1
    # via openai
pydantic-core==2.18.2
    # via pydantic
python-dotenv==1.0.1
regex==2024.4.16
    # via tiktoken
requests==2.31.0
    # via
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via omnivoreql
setuptools==69.5.1
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tiktoken==0.6.0
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
wheel==0.43.0
yarl==1.9.4
    # via gql
